requirements.txt
streamlit==1.39.0
streamlit-drawable-canvas
supervision>=0.19.0
inference-gpu[yolo-world]==0.62.5
opencv-python-headless
python-dotenv
Pillow
